README.md: Clarify python version 3.6 as the requirement
It was reported that the bluepy-scratch-link does not run on Ubuntu 16.04
which has python version 3.5. The script stopped with an error about
f-format string, which was introduce with python 3.6.

Even removing f-format string from the script, bluepy-scratch-link still
requires python 3.6 because websockets module requires it. Clarify that
python version requirement to avoid confusion.

Signed-off-by: Shin'ichiro Kawasaki <kawasaki@juno.dti.ne.jp>