gencert.sh: Automate certificate addition to FireFox and Chrome
As of today, bluepy-scratch-link users need to do special action to allow
local server certificates. This is trouble some and James Le Cuirot
suggested to automate the action with certutil tools

To avoid the user action, check if NSS DB of FireFox or Chrome exists. If
NSS DBs exist, add the certificate to those DBs.

Signed-off-by: Shin'ichiro Kawasaki <kawasaki@juno.dti.ne.jp>