Added self.deviceName in BLESession
In logger.* the device was print out as "<bluepy.btle.Peripheral object at 0x************>" which is not very useful to the user.
The deviceName variable can be also useful if the name of the device is needed and it does not matter if it is the short name or the long name of the device.
The shorter device name will be used if the long name is not defined